accessible/tests/mochitest/hittest/test_shadowroot_subframe.html

<!DOCTYPE HTML>
<html>
<head>
  <title>ShadowRoot hit tests</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="application/javascript" src="../common.js"></script>
  <script type="application/javascript" src="../layout.js"></script>

  <script type="application/javascript">
    let SimpleTest = window.parent.SimpleTest;
    let ok = window.parent.ok;
    let is = window.parent.is;

    function doTest() {
      var componentAcc = getAccessible("component1");
      testChildAtPoint(componentAcc, 1, 1, componentAcc.firstChild,
                       componentAcc.firstChild);

      componentAcc = getAccessible("component2");
      testChildAtPoint(componentAcc, 1, 1, componentAcc.firstChild,
                       componentAcc.firstChild);
      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>

</head>
<body>
  <div role="group" class="components" id="component1" style="display: inline-block;">
  <!--
    <div role="button" id="component-child"
         style="width: 100px; height: 100px; background-color: pink;">
    </div>
  -->
  </div>
  <div role="group" class="components"  id="component2" style="display: inline-block;">
  <!--
    <button>Hello world</button>
  -->
  </div>
  <script>
    // This routine adds the comment children of each 'component' to its
    // shadow root.
    var components = document.querySelectorAll(".components");
    for (var i = 0; i < components.length; i++) {
      var component = components[i];
      var shadow = component.attachShadow({mode: "open"});
      for (var child = component.firstChild; child; child = child.nextSibling) {
        if (child.nodeType === 8)
          // eslint-disable-next-line no-unsanitized/property
          shadow.innerHTML = child.data;
      }
    }
  </script>
</body>
</html>
